refactor:  migrate lib/crudService to ts
